move schema injection into utils
